Time: 57 ms (49.5%), Space: 107.8 MB (84.5%) - LeetHub

+++ 3867-sum-of-gcd-of-formed-pairs/3867-sum-of-gcd-of-formed-pairs.java
@@ -12,7 +12,7 @@ class Solution {
             if(nums[i]>x){
                 x = nums[i];
             }
-            prefixGCD[i] = (int)gcd(x,nums[i]);
+            prefixGCD[i] = gcd(x,nums[i]);
         }
         Arrays.sort(prefixGCD);
         int i = 0;
